requirements.txt
fastapi==0.121.0
pydantic==2.12.3
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.51.0
